Attempt to fix long standing router issue (#83) that causes '.' to be disallowed within named placeholders in routes.
The proposed solution is to make the named placeholder regexp fragment non-greedy so that it won't consume an optional format. This allows the removal of '.' from the negative character class.

The only case this doesn't seem to cover is a named placeholder that can contain '.' followed by an optional format, but disambiguating between those two cases seems impossible in the general case and shouldn't be relied upon anyway.