Add `watch` option to staticProvider
If the `watch` option value is truthy, fs.watchFile will be used on
any file that gets cached, and any change will invalidate the cache
and unwatch the file. (The option does nothing unless caching is
enabled).

On Linux systems, watchFile uses inotify, and should be reasonably
performant. On other systems, it will periodically stat() the file in
question, and thus is undesirable for production.