Fixed potential security issue, store files in req.files. Closes #431
this oversight as described in the issue allows
for an attacker to essentially fake file paths.

this depends on how you are using the module, however
to keep things safer files will now be stored in req.files
and not req.body, much like signed / unsigned cookies are
stored in separate objects to prevent a similar issue.