add `secret` option to `session` middleware
This decouples `session` and `cookieParser` middleware by not requiring
the `cookieParser` middleware to do signed cookies for the session
middleware. This separates middleware responsibility and allows each
middleware to be a bit more self contained and implement signing within
session as needed.

Eventually this will allow the `cookieParser` middleware to just do the
basic functions of cookie parsing instead of trying to do too much.