add `secret` option to `cookieSession` middleware
This decouples `cookieParser` and `cookieSession` by not requiring the
`cookieParser` to handle the `cookieSession` signed cookies. This allows
the `cookieParser` middleware to be simpler by not having to perform
hashing. It also keeps the cookie session functionality contained
withing the `cookieSession` middleware.

cookieHashes has been removed as it is no longer needed. The hashing was
only done for the `cookieSession` middleware