prevent options object from being passed to qs as delim
As far as I can tell, versions of qs prior to 1.0.0, there has never
been an options object parameter to qs.parse, however connect has
supported passing such an object to parse since a73dd2b. In qs 1.2.0,
which the 2.x branch of connect is currently dependent on, parse takes 3
parameters, and if the second parameter is not a number, it is treated
as the third parameter, the delimiter. As a result, you'll end up with
the object as the delimiter, which isn't good.

What I've done here is remove the ability to pass the options object to
qs.parse. If it's there, it will just be ignored. It was never actually
being used anyway.